requirements.txt
streamlit
openai
supabase
langchain
langchain-community
langchain-text-splitters
pypdf
python-dotenv
tqdm
termcolor
pydantic
